requirements.txt
# populate with required packages for project
ConfigArgParse==0.14.0
cycler==0.10.0
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.1
munkres==1.1.2
numpy==1.16.4
pandas==0.24.2
Pillow==6.1.0
protobuf==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytorch-ignite==0.2.0
pytz==2019.1
PyYAML==5.1.1
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
tensorboardX==1.8
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
yacs==0.1.6
